% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{annotate_peaks}
\alias{annotate_peaks}
\title{annotate_peaks}
\usage{
annotate_peaks(
  peakfile,
  gtffile,
  genome = "hg19",
  fiveP = 1000,
  threeP = 1000,
  simple = FALSE,
  RNA = TRUE
)
}
\arguments{
\item{peakfile, }{a string denoting the peak file name, only .bed format is allowed}

\item{gtffile, }{a string denoting the path to the gtf or gff3 file}

\item{fiveP, }{extension out of the 5' boundary of genes for defining promoter: -fiveP TSS +100}

\item{threeP, }{extension out of the 3' boundary of genes for defining termination region: -100 TTS +threeP}

\item{simple, }{a boolean object indicating whether 5'UTR and 3'UTR are annotated in the gtffile}

\item{RNA, }{a boolean object indicating whether only peaks in transcripts should be considered for pie chart plot}

\item{peaklabel, }{a string denoting short labels to appeare in the plot}
}
\description{
Produce a table of genes targeted by peaks, and generate plots for target gene types, peak distribution in genomic features
}
