% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDataMatrix.R
\name{aov_TukeyHSD}
\alias{aov_TukeyHSD}
\title{Perform one-way ANOVA and post hoc TukeyHSD tests}
\usage{
aov_TukeyHSD(df, xc = "Group", yc = "Intensity", op = NULL, verbose = FALSE)
}
\arguments{
\item{df}{a dataframe with c("Intensity", "Group") in column names}

\item{xc}{a string denoting column name for grouping}

\item{yc}{a string denoting column name for numeric data to be plotted}

\item{op}{output prefix for statistical analysis results}

\item{verbose}{logical, to indicate whether a file should be produced to save the test results}
}
\value{
a list of two matrices, the first is the summary of ANOVA test and the second is the output of TukeyHSD tests
}
\description{
This is a helper function for performing one-way ANOVA analysis and post hoc Tukey's Honest Significant Differences tests
}
\note{
used in \code{plot_reference_locus}
}
\author{
Shuye Pu
}
