% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{draw_boxplot_logy}
\alias{draw_boxplot_logy}
\title{Plot boxplot with log scale y-axis}
\usage{
draw_boxplot_logy(
  stat_df,
  xc = "Feature",
  yc = "Intensity",
  fc = xc,
  comp = list(c(1, 2)),
  stats = "wilcox.test",
  Xlab = xc,
  Ylab = yc,
  logy = "linear",
  nf = 1,
  pc = 1
)
}
\arguments{
\item{stat_df}{a dataframe with column names c(xc, yc)}

\item{xc}{a string denoting column name for grouping}

\item{yc}{a string denoting column name for numeric data to be plotted}

\item{fc}{a string denoting column name for sub-grouping}

\item{comp}{a list of vectors denoting pair-wise comparisons to be performed between groups}

\item{stats}{the name of pair-wise statistical tests, like t.test or wilcox.test}

\item{Ylab}{a string for y-axis label}

\item{logy}{character string in c(linear, logy, log2, log10, wo), "wo" = without outliers}

\item{nf}{a integer normalizing factor for correct count of observations when the data table is produced by pivot_longer, equals to the number of columns in cols argument of pivot_longer}

\item{pc}{pseudo-count for log transformation}
}
\value{
a ggplot object
}
\description{
Plot boxplot with linear or log scale y-axis, with p-value significance levels displayed
}
\note{
used by \code{plot_reference_locus}, \code{plot_reference_locus_with_random}
}
\author{
Shuye Pu
}
