% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{draw_combo_plot}
\alias{draw_combo_plot}
\title{Make combo plot for statistics plots}
\usage{
draw_combo_plot(
  stat_df,
  xc = "Feature",
  yc = "Intensity",
  comp = list(c(1, 2)),
  Xlab = xc,
  Ylab = yc,
  fc = xc,
  Ylim = NULL,
  nf = 1
)
}
\arguments{
\item{stat_df}{a dataframe with column names c(xc, yc)}

\item{xc}{a string denoting column name for grouping}

\item{yc}{a string denoting column name for numeric data to be plotted}

\item{comp}{a list of vectors denoting pair-wise comparisons to be performed between groups}

\item{Xlab}{a string for x-axis label}

\item{Ylab}{a string for y-axis label}

\item{fc}{a string denoting column name for sub-grouping based on an additional factor}

\item{Ylim}{a numeric vector of two elements, defining custom limits of y-axis}

\item{nf}{a integer normalizing factor for correct count of observations when the data table has two factors, such as those produced by pivot_longer, equals to the number of factors}
}
\value{
a ggplot object
}
\description{
Place violin plot, boxplot without outliers, mean+se barplot and quantile plot on the same page
}
\author{
Shuye Pu
}
