% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{draw_region_name}
\alias{draw_region_name}
\title{Plot genomic region names}
\usage{
draw_region_name(featureNames, scaled_bins, xmax)
}
\arguments{
\item{featureNames}{a string vector giving names of sub-regions}

\item{scaled_bins}{a vector on integers denoting the length of each sub-region}

\item{xmax}{an integer denoting the left most boundary}
}
\value{
a ggplot object
}
\description{
Plot sub-region labels under the landmark
}
\note{
used by \code{plot_3parts_metagene}, \code{plot_5parts_metagene}, \code{plot_reference_region}
}
\author{
Shuye Pu
}
