% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{draw_stacked_plot}
\alias{draw_stacked_plot}
\title{draw stacked plot}
\usage{
draw_stacked_plot(plot_list, heatmap_list)
}
\arguments{
\item{plot_list}{a list of profile plots}

\item{heatmap_list}{a list of heatmaps}
}
\value{
a null value
}
\description{
Plot profile on top of heatmap, and align feature labels.
}
\note{
used by \code{\link{plot_locus}}, \code{\link{plot_5parts_metagene}},
   \code{\link{plot_region}}
}
\author{
Shuye Pu
}
