% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{draw_stacked_profile}
\alias{draw_stacked_profile}
\title{Plot signal profile around start, center, and end of genomic regions}
\usage{
draw_stacked_profile(
  plot_df,
  xc = "Position",
  yc = "Intensity",
  cn = "Query",
  ext = c(0, 0, 0, 0),
  hl = c(0, 0, 0, 0),
  atitle = "title",
  insert = 0,
  Ylab = "Signal Intensity",
  shade = FALSE
)
}
\arguments{
\item{plot_df}{a dataframe with column names c(xc, yc, cn, "Interval", "lower", "upper")}

\item{xc}{a string denoting column name for values on x-axis}

\item{yc}{a string denoting column name for numeric data to be plotted}

\item{cn}{a string denoting column name for grouping}

\item{ext}{a vector of 4 integers denoting upstream and downstream extension around start and end}

\item{hl}{a vector of 4 integers defining upstream and downstream boundaries of the rectangle for start and end}

\item{atitle}{a string for the title of the plot}

\item{insert}{a integer denoting the width of the center region}

\item{Ylab}{a string for y-axis label}

\item{shade}{logical indicating whether to place a shaded rectangle around the point of interest}
}
\value{
a ggplot object
}
\description{
Plot lines with standard error as the error band, also plots number of regions having non-zero signals
}
\note{
used by \code{plot_start_end_feature}, \code{plot_start_end_reference_region}
}
\author{
Shuye Pu
}
