% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{effectiveSize}
\alias{effectiveSize}
\title{effectiveSize}
\usage{
effectiveSize(outlist, outRle, genome = "hg19")
}
\arguments{
\item{outlist, }{a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type,
'weight' is the name of the metadata column}

\item{outRle, }{a boolean object indicating whether the output is a list of RleList objects or GRanges objects}

\item{genome, }{a string denoting the genome version}
}
\value{
a list object with four elements, with the 'size' element modified.
}
\description{
effectiveSize
}
\author{
Shuye Pu
}
