% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\docType{data}
\name{extdata}
\alias{extdata}
\title{Toy data for examples and testing}
\value{
These files will be read by handle_input function, which will 
     produce GRanges object or RleList object, depending on import parameters
}
\description{
The data files in the extdata directory contain data for next 
generation sequencing read alignments, MACS2 peaks and gene annotation, which 
are used to test the package and generate plots in the package vignettes. 
Except for the gtf file, all other files are derived from experimental data 
produced in-house. To meet the package file size limit, all data are 
restricted to chr19:58000-507000 of the human genome version hg19. Details 
for each file are as follows.
}
\details{
\itemize{
 \item "gencode.v19.annotation_chr19.gtf" is an excerpt of the gene 
     annotation file downloaded from 
     \url{https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_human/release_19/gencode.v19.annotation.gtf.gz}. 
     "gencode.v19.annotation_chr19.gtf.granges.rds" is a GRanges object as a 
     result of importing the gtf file.
 \item "chip_treat_chr19.bam(.bai)" and "chip_input_chr19.bam(.bai)" are 
     paired-end read alignment data from ChIPseq experiments.
 \item "treat_chr19.bam(.bai)" and "input_chr19.bam(.bai)" are single-end 
     read alignment data from iCLIP experiments.
 \item "test_wig_chr19_+(-).wig", "test_wig_chr19_+(-).bw" are iCLIP 
     alignment data in WIG and BIGWIG format, respectively; '+' and '-' 
     represent forward and reverse strand, respectively. 
 \item "test_clip_peak_chr19.bed" contains strand-specific iCLIP peak in BED 
     format. 
 \item "test_chip_peak_chr19.bed" and "test_chip_peak_chr19.narrowPeak" 
     contain ChIPseq peaks generated with MACS2, in summit peak and narrow 
     peak format, respectively. 
 \item "test_file1.txt", "test_file2.txt", "test_file3.txt" and 
     "test_file4.txt" are tab-delimited text files,  each contains various 
     human gene names in different columns. 
}
}
\author{
Shuye Pu
}
\keyword{datasets}
