% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{format_genomic_coordinates}
\alias{format_genomic_coordinates}
\title{Format genomic coordinates in GRanges or GRrangesList as strings used in igv}
\usage{
format_genomic_coordinates(x)
}
\arguments{
\item{x}{a GRanges or GRangesList object}
}
\value{
a vector of strings in the format of 'chr:start-end(strand)'
}
\description{
This function takes a GRanges or GRangesList object, and transform each range into a string
}
\examples{
gr1 <- GenomicRanges::GRanges("chr2", IRanges::IRanges(3, 6))
gr2 <- GenomicRanges::GRanges(c("chr1", "chr1"), IRanges::IRanges(c(7,13), width=3),
 strand=c("+", "-"))
gr3 <- GenomicRanges::GRanges(c("chr1", "chr2"), IRanges::IRanges(c(1, 4), c(3, 9)),
 strand="-")

grl <- GenomicRanges::GRangesList(gr1= gr1, gr2=gr2, gr3=gr3)
grl

out <- format_genomic_coordinates(grl)
cat(out)

}
\author{
Shuye Pu
}
