% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{gtf_to_bed_longest_tx}
\alias{gtf_to_bed_longest_tx}
\title{gtf_to_bed_longest_tx
Make bed and bed 12 files from gtf file for protein coding genes. For "utr3", "utr5", "cds" and "transcript",  the output granges object
gives start and end of the entire feature, while the grangslist object gives start and end of each exonic segment. For "exon"
and "intron", the output granges object gives start and end of each feature, while the grangslist object also provides
transcript ID in addition to these information. For "gene", the granges object and the grangeslist object are equivalent.
If 'longest'=TRUE, only the longest transcript of each gene is considered, the exons, introns, 3'utr and 5'utr are limited to
the longest transcript of protein-coding genes.}
\usage{
gtf_to_bed_longest_tx(
  txdb,
  featureName,
  featureSource = NULL,
  export = FALSE,
  longest = TRUE
)
}
\arguments{
\item{txdb, }{a TxDb object defined in GenomicFeatures package}

\item{featureSource, }{the name of the gtf/gff3 file or the online database from which txdb is derived, used as name of output file}

\item{export, }{a boolean object to indicate if the bed file should be produced}

\item{longest, }{a boolean object to indicate whether the output should be limited to the longest transcript of each gene}

\item{feaureNmae, }{one of the gene feature in c("utr3", "utr5", "cds", "intron", "exon", "transcript", "gene")}
}
\value{
a list of two objects, the first is a GRanges object, the second is a GRangesList object
}
\description{
gtf_to_bed_longest_tx
Make bed and bed 12 files from gtf file for protein coding genes. For "utr3", "utr5", "cds" and "transcript",  the output granges object
gives start and end of the entire feature, while the grangslist object gives start and end of each exonic segment. For "exon"
and "intron", the output granges object gives start and end of each feature, while the grangslist object also provides
transcript ID in addition to these information. For "gene", the granges object and the grangeslist object are equivalent.
If 'longest'=TRUE, only the longest transcript of each gene is considered, the exons, introns, 3'utr and 5'utr are limited to
the longest transcript of protein-coding genes.
}
\examples{
txdb <- makeTxDbFromEnsembl("Saccharomyces cerevisiae", server="useastdb.ensembl.org")
output <- gtf_to_bed_longest_tx(txdb, featureName="utr3", featureSource="Ensembl", export=F, longest=T)
}
