% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{handle_bam}
\alias{handle_bam}
\title{handle_bam}
\usage{
handle_bam(
  inputFile,
  CLIP_reads = FALSE,
  fix_width = 0,
  outRle = TRUE,
  genome = "hg19"
)
}
\arguments{
\item{inputFile, }{a string denoting the bed file name}

\item{CLIP_reads, }{a boolean object to indicate if the reads should be shifted to their -1 position at the 5' end and shrink to length 1}

\item{fix_width, }{an integer defines how long should the reads should be extended to}

\item{outRle, }{a boolean object indicating whether the output should be a list of RleList objects or GRanges objects}

\item{genome, }{a string denoting the genome version}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type,
'weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
handle_bam
}
\details{
The reads are filtered use mapq score >= 10 by default, only mapped reads are counted towards library size.
}
\author{
Shuye Pu
}
