% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{handle_bed}
\alias{handle_bed}
\title{handle_bed}
\usage{
handle_bed(
  inputFile,
  fix_width = 0,
  useScore = FALSE,
  outRle = TRUE,
  fixPoint = "center",
  genome = "hg19"
)
}
\arguments{
\item{inputFile, }{a string denoting the bed file name}

\item{fix_width, }{an integer defines how long should the peaks should be extended to}

\item{useScore, }{a boolean object indicating whether the 'score' column of the bed file should be used in the output data structure}

\item{outRle, }{a boolean object indicating whether the output should be a list of RleList objects or GRanges objects}

\item{genome, }{a string denoting the genome version}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type,
'weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
handle_bed
}
\details{
when useScore is TRUE, the score column of the bed file will be used in the metadata column 'score' of the GRanges object,
or the 'Values' field of the RleList object. Otherwise the value 1 will be used instead. When the intended use of the input bed is a
reference feature, both useScore and outRle should be set to FALSE.
}
\author{
Shuye Pu
}
