% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{handle_input}
\alias{handle_input}
\title{handle_input}
\usage{
handle_input(
  inputFiles,
  CLIP_reads = FALSE,
  fix_width = 0,
  useScore = FALSE,
  outRle = TRUE,
  useSizeFactor = FALSE,
  genome = "hg19"
)
}
\arguments{
\item{inputFiles, }{a vector of strings denoting file names}

\item{CLIP_reads, }{a boolean object to indicate if the bam reads should be shifted to the -1 position at the 5' of the reads}

\item{fix_width, }{an integer defines how long should the reads should be extended to}

\item{useScore, }{a boolean object indicating whether the 'score' column of the bed file should be used in the output data structure}

\item{outRle, }{a boolean object indicating whether the output should be a list of RleList objects or GRanges objects}

\item{genome, }{a string denoting the genome version}

\item{norm, }{a boolean object indicating whether the library size should be adjusted with a size factor}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type,
'weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
handle_input
}
\author{
Shuye Pu
}
