% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleFeatures.R
\name{makeSubTxDbFromGtf}
\alias{makeSubTxDbFromGtf}
\title{Make TxDb object from a GTF file for a subset of genes.}
\usage{
makeSubTxDbFromGtf(gtfFile, geneList, geneCol = 1)
}
\arguments{
\item{geneList}{path to a line-delimited text file with one gene name on each line, or a character vector of gene names}

\item{gtfFfile}{path to a GTF file}
}
\value{
a TxDb object
}
\description{
Make a partial TxDb object given a GTF file and a list of gene names in a file or in a character vector.
}
\author{
Shuye Pu
}
