% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{overlap_pair}
\alias{overlap_pair}
\title{Plot two-sets Venn diagram}
\usage{
overlap_pair(apair, overlap_fun)
}
\arguments{
\item{apair}{a list of two vectors}

\item{overlap_fun}{the name of the function that defines overlap, depending on the type of object in the vectors.}
}
\description{
This is a helper function for Venn diagram plot. A Venn diagram is plotted as output.
}
\author{
Shuye Pu
}
