% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{overlap_quad}
\alias{overlap_quad}
\title{Plot four-sets Venn diagram}
\usage{
overlap_quad(aquad, overlap_fun)
}
\arguments{
\item{aquad}{a list of four vectors}

\item{overlap_fun}{the name of the function that defines overlap}
}
\description{
This is a helper function for Venn diagram plot. A Venn diagram is plotted as output.
}
\author{
Shuye Pu
}
