% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{overlap_triple}
\alias{overlap_triple}
\title{Plot three-sets Venn diagram}
\usage{
overlap_triple(atriple, overlap_fun)
}
\arguments{
\item{atriple}{a list of three vectors}

\item{overlap_fun}{the name of the function that defines overlap}
}
\description{
This is a helper function for Venn diagram plot. A Venn diagram is plotted as output.
}
\author{
Shuye Pu
}
