% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{parallel_apply_scoreMatrixBin}
\alias{parallel_apply_scoreMatrixBin}
\title{parallel_apply_scoreMatrixBin}
\usage{
parallel_apply_scoreMatrixBin(
  windowRs,
  queryRegions,
  bin_num,
  bin_op,
  weight_col,
  stranded
)
}
\arguments{
\item{windowRs, }{a list of GRangesList objects.}

\item{queryRegions, }{a RleList object or Granges object providing input for the 'target' parameter of the scoreMatrixBin method}

\item{bin_num, }{number of bins the windows should be divided into}

\item{bin_op, }{operation on the signals in a bin, a string in c("mean", "max", "min", "median", "sum") is accepted.}

\item{weight_col, }{if the queryRegions is a GRanges object, a numeric column in meta data part can be used as weights.}

\item{stranded, }{a boolean object to indicate if the strand of the windows should be considered to determine upstream and downstream}
}
\value{
a list of numeric matrices
}
\description{
Function for parallel computation of scoreMatrixBin. The 'windows' parameter of the scoreMatrixBin method is provided as a list of
GRangesList objects, allowing simultaneous processing of multiple windows objects.
}
