% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{parallel_binnedAverage}
\alias{parallel_binnedAverage}
\title{Parallel execution of binnedAverage}
\usage{
parallel_binnedAverage(Rle_list, tileBins, nc = 2)
}
\arguments{
\item{Rle_list}{a list of RleList objects.}

\item{tileBins}{a GRanges object of tiled genome}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a list of numeric vectors
}
\description{
Function for parallel computation of binnedAverage function in the GenomicRanges package
}
\author{
Shuye Pu
}
