% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{parallel_countOverlaps}
\alias{parallel_countOverlaps}
\title{Parallel execution of countOverlaps}
\usage{
parallel_countOverlaps(grange_list, tileBins, nc = 2, switch = FALSE)
}
\arguments{
\item{grange_list}{a list of GRanges objects.}

\item{tileBins}{a GRanges object of tiled genome}

\item{nc}{integer, number of cores for parallel processing}

\item{switch}{logical, switch the order of query and feature}
}
\value{
a list of numeric vectors
}
\description{
Function for parallel computation of countOverlaps function in the GenomicRanges package
}
\author{
Shuye Pu
}
