% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{parallel_feature_coordinates}
\alias{parallel_feature_coordinates}
\title{Parallel execution of get_genomic_feature_coordinates}
\usage{
parallel_feature_coordinates(
  txdb,
  featureNames,
  longest = TRUE,
  protein_coding = TRUE,
  nc = 2
)
}
\arguments{
\item{txdb}{a TxDb object}

\item{featureNames}{a vector of gene features in c("utr3", "utr5", "cds", "intron", "exon", "transcript", "gene")}

\item{longest}{logical, indicating whether the output should be limited to the longest transcript of each gene}

\item{protein_coding}{logical, indicating whether to limit to protein_coding genes}

\item{nc}{number of cores to use}
}
\value{
a list of lists(Granges, GRangesList)
}
\description{
Get genomic coordinates for multiple features at once
}
\note{
not working because txdb cannot be exported to nodes, need to find a solution
}
\author{
Shuye Pu
}
\keyword{internal}
