% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_start_end_feature}
\alias{plot_start_end_feature}
\title{plot_start_end_feature}
\usage{
plot_start_end_feature(
  queryfiles,
  inputfiles = NULL,
  txdb,
  featureName,
  CLIP_reads = F,
  binsize = 10,
  fix_width = 0,
  longest = T,
  ext = c(-500, 200, -200, 500),
  hl = c(-50, 50, -50, 50),
  randomize = FALSE,
  stranded = T,
  norm = F,
  scale = F,
  smo = F,
  heatmap = F,
  rm.outlier = F,
  genome = "hg19",
  outPrefix = "plots",
  useScore = FALSE,
  useSizeFactor = FALSE
)
}
\arguments{
\item{queryfiles, }{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{txdb, }{a TxDb object defined in GenomicFeatures package}

\item{CLIP_reads, }{a boolean object to indicate if the bam reads should be shifted to the -1 positon at the 5' of the reads}

\item{binsize, }{an integer defines bin size for intensity calculation}

\item{fix_width, }{an integer defines how long should the reads should be extended to}

\item{longest, }{a boolean object to indicate whether the output should be limited to the longest transcript of each gene}

\item{ext, }{a vector of four integers defining upstream and downstream boundaries of the plot window, flanking the start and end of features}

\item{hl, }{a vector of four integers defining upstream and downstream boundaries of the highlight window, flanking the start and end of features}

\item{randomize, }{a boolean indicate if randomized feature should generated and used as a contrast to the real feature}

\item{stranded, }{a boolean indicate whether the strand of the feature should be considered}

\item{norm, }{an boolean indicates if the intensities should be normalized by the sample library sizes}

\item{scale, }{a boolean indicates whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{smo, }{a boolean indicates whether the line should smoothed with a spline smoothing algorithm}

\item{heatmap, }{a boolean indicates whether a heatmap of the score matrix should be generated}

\item{rm.outlier, }{a boolean indicates whether a row with abnormally high values in the score matrix should be removed}

\item{genome, }{genome of the features, the program is mostly concerned with human, if non-human genome is used, certain features may not work}

\item{outPrefix, }{a character string specifying output file prefix for plots (outPrefix.pdf)}

\item{querylabels, }{a vector of charactor strings serving as short labels of the queryfiles, will be used as sample labels in the plots}

\item{feaureNmae, }{one of the gene feature in c("utr3", "utr5", "cds", "intron", "exon", "transcript", "gene")}
}
\value{
a list of two objects, the first is a GRanges object, the second is a GRangesList object
}
\description{
plot_start_end_feature
}
\section{Functions}{
\itemize{
\item \code{plot_start_end_feature}: reads or peak signal intensity of samples in the query files around stat and end of genomic features. The upstream and downstream windows
can be given separately, within the window, a smaller window can be defined to test statistical difference between the signal intensity
among samples. A line plot and a boxplot are displayed side by side for both start and end of feature.
}}

\examples{
txdb <- makeTxDbFromEnsembl("Homo sapiens", release=75)
queryfiles <- "YTHDF2.merged.thUni.bam"
querylabels <- "YTHDF2_merged"
inputfiles <- "Input_YTHDF2.thUni.bam"
inputlabels <- "YTHDF2_Input"
ext <- c(-500, 200, -200, 500)
hl <- c(-50, 50, -50, 50)
op <- "ratio_YTHDF2_merged_at_intron_boundary"

plot_start_end_feature(queryfiles=queryfiles, querylabels=querylabels, inputfiles=inputfiles, inputlabels=inputlabels, txdb=txdb, featureName="intron", CLIP_read=F, binsize=10, fix_width=0,
longest=T, ext=ext, hl=hl, randomize=T, stranded=T, norm=F, scale=F, smo=F, heatmap=F, rm.outlier=F, genome="hg19", outPrefix=op)

}
