% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_bam_correlation}
\alias{plot_bam_correlation}
\title{plot_bam_correlation}
\usage{
plot_bam_correlation(
  bamfiles,
  binsize = 1e+06,
  outPrefix = "Bam_correlation",
  genome = "hg19"
)
}
\arguments{
\item{bamfiles, }{a vector of strings denoting file names}

\item{binsize, }{an integer denoting the tile width for tiling the genome, default 1000000}

\item{outPrefix, }{a string denoting output file name in pdf format}

\item{genome, }{a string denoting the genome version}

\item{bamlabels, }{a vector of strings denoting short labels to appeare in the plot}
}
\description{
plot_bam_correlation
}
\examples{
bamfiles <- queryfiles
genome <- "hg19"
binsize <- 10000
outPrefix="Bam_correlation"
}
