% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_named_list}
\alias{plot_named_list}
\title{plot a named list as a figure}
\usage{
plot_named_list(params)
}
\arguments{
\item{params}{a list produced by as.list(environment()), with names being the arguments and values being the runtime values when the function is called,}
}
\value{
a ggplot object
}
\description{
This is a helper function for displaying function arguments for a plotting function
}
\author{
Shuye Pu
}
\keyword{internal}
