% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_reference_locus_with_random}
\alias{plot_reference_locus_with_random}
\title{plot_reference_locus_random
Plot reads or peak signal intensity of samples in the query files around reference locus defined in the centerfiles. The upstream and downstream windows flanking genes
can be given separately, a smaller window can be defined to allow statistical comparisons between reference and random loci.}
\usage{
plot_reference_locus_with_random(
  queryfiles,
  centerfiles,
  txdb,
  ext = c(0, 0),
  hl = c(0, 0),
  shade = F,
  useSizeFactor = FALSE,
  smo = FALSE,
  CLIP_reads = FALSE,
  fix_width = 0,
  norm = FALSE,
  binsize = 10,
  refPoint = "center",
  Xlab = "Center",
  inputfiles = NULL,
  stranded = TRUE,
  heatmap = FALSE,
  scale = FALSE,
  outPrefix = NULL,
  genome = "hg19",
  rm.outlier = F,
  n_random = 1,
  stats.method = "wilcox.test",
  useScore = FALSE
)
}
\arguments{
\item{queryfiles, }{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{centerfiles, }{a vector of reference file names. The file should be .bed format only}

\item{ext, }{a vector of two integers defining upstream and downstream boundaries of the plot window, flanking the reference locus}

\item{hl, }{a vector of two integers defining upstream and downstream boundaries of the highlight window, flanking the reference locus}

\item{smo, }{a boolean indicates whether the line should smoothed with a spline smoothing algorithm}

\item{CLIP_reads, }{a boolean object to indicate if the bam reads should be shifted to the -1 position at the 5' of the reads}

\item{fix_width, }{an integer defines how long should the reads should be extended to}

\item{norm, }{an boolean indicates if the intensities should be normalized by the sample library sizes}

\item{refPoint, }{a sting in c("start", "center", "end")}

\item{Xlab, }{a string denotes the label on x-axis}

\item{inputfiles, }{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{stranded, }{a boolean indicate whether the strand of the feature should be considered}

\item{heatmap, }{a boolean indicates whether a heatmap of the score matrix should be generated}

\item{scale, }{a boolean indicates whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{outPrefix, }{a character string specifying output file prefix for plots (outPrefix.pdf)}

\item{genome, }{genome of the features, the program is mostly concerned with human, if non-human genome is used, certain features may not work}

\item{rm.outlier, }{a boolean indicates whether a row with abnormally high values in the score matrix should be removed}

\item{n_random, }{an integer denotes the number of randomization should be formed}

\item{stats.method, }{a string in c("wilcox.test", "t.test"), for pair-wise groups comparisons}

\item{querylabels, }{a vector of character strings serving as short labels of the queryfiles, will be used as sample labels in the plots}

\item{centerlabels, }{a vector of character strings serving as short labels of the centerfiles, will be used as sample labels in the plots}

\item{inputlabels, }{a vector of character strings serving as short labels of the inputfiles, will be used as input labels in the plots}

\item{longest, }{a boolean object to indicate whether the output should be limited to the longest transcript of each gene}
}
\value{
a dataframe containing the data used for plotting
}
\description{
plot_reference_locus_random
Plot reads or peak signal intensity of samples in the query files around reference locus defined in the centerfiles. The upstream and downstream windows flanking genes
can be given separately, a smaller window can be defined to allow statistical comparisons between reference and random loci.
}
