% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_reference_region}
\alias{plot_reference_region}
\title{plot_reference_region
Plot reads or peak signal intensity of samples in the query files inside regions defined in the centerfiles. The upstream and downstream windows flanking genes
can be given separately, the parameter 'meta' controls if gene or metagene plots are generated.}
\usage{
plot_reference_region(
  queryfiles,
  centerfiles,
  inputfiles = NULL,
  nbins = 100,
  norm = FALSE,
  useSizeFactor = FALSE,
  fix_width = 0,
  CLIP_reads = FALSE,
  scale = FALSE,
  heatmap = FALSE,
  longest = TRUE,
  useScore = FALSE,
  fiveP = 1000,
  threeP = 1000,
  smo = FALSE,
  stranded = TRUE,
  outPrefix = "plots",
  genome = "hg19",
  rm.outlier
)
}
\arguments{
\item{queryfiles, }{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{centerfiles, }{a vector of reference file names. The file should be .bed format only}

\item{inputfiles, }{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{nbins, }{an integer defines the total number of bins}

\item{norm, }{an boolean indicates if the intensities should be normalized by the sample library sizes}

\item{fix_width, }{an integer defines how long should the reads should be extended to}

\item{CLIP_reads, }{a boolean object to indicate if the bam reads should be shifted to the -1 position at the 5' of the reads}

\item{scale, }{a boolean indicates whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{heatmap, }{a boolean indicates whether a heatmap of the score matrix should be generated}

\item{longest, }{a boolean object to indicate whether the output should be limited to the longest transcript of each gene}

\item{fiveP, }{extension out of the 5' boundary of gene}

\item{threeP, }{extension out of the 3' boundary of gene}

\item{smo, }{a boolean indicates whether the line should smoothed with a spline smoothing algorithm}

\item{stranded, }{a boolean indicate whether the strand of the feature should be considered}

\item{outPrefix, }{a character string specifying output file prefix for plots (outPrefix.pdf)}

\item{genome, }{genome of the features, the program is mostly concerned with human, if non-human genome is used, certain features may not work}

\item{rm.outlier, }{a boolean indicates whether a row with abnormally high values in the score matrix should be removed}

\item{querylabels, }{a vector of character strings serving as short labels of the queryfiles, will be used as sample labels in the plots}

\item{centerlabels, }{a vector of character strings serving as short labels of the centerfiles, will be used as sample labels in the plots}

\item{inputlabels, }{a vector of character strings serving as short labels of the inputfiles, will be used as input labels in the plots}
}
\value{
a dataframe containing the data used for plotting
}
\description{
plot_reference_region
Plot reads or peak signal intensity of samples in the query files inside regions defined in the centerfiles. The upstream and downstream windows flanking genes
can be given separately, the parameter 'meta' controls if gene or metagene plots are generated.
}
