% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{process_scoreMatrix}
\alias{process_scoreMatrix}
\title{process_scoreMatrix}
\usage{
process_scoreMatrix(
  fullmatrix,
  libsize = 1e+06,
  norm = F,
  scale = F,
  heatmap = F,
  rm.outlier = F
)
}
\arguments{
\item{fullmatrix, }{a numeric matrix, with bins in columns and genomic windows in rows}

\item{libsize, }{sample library size in terms of read counts}

\item{norm, }{a boolean object indicating whether the matrix should be normalized to reads per million (RPM)}

\item{scale, }{a boolean indicates whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{heatmap, }{a boolean indicates whether a heatmap of the score matrix should be generated}

\item{rm.outlier, }{a boolean indicates whether a row with abnormally high values in the score matrix should be removed}
}
\value{
a numeric matrix
}
\description{
This is a helper function for manipulate the score matrix produced by ScoreMatrix or ScoreMatrinBin functions defined in the 'genomation' package.
}
