% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDataMatrix.R
\name{rank_rows}
\alias{rank_rows}
\title{Rank rows of a matrix based on user input}
\usage{
rank_rows(fullmatrix, ranking = "Hierarchical")
}
\arguments{
\item{fullmatrix}{a numeric matrix}

\item{ranking}{a string in c("Sum", "Max", "Hierarchical", "None")}
}
\value{
a numerical matrix
}
\description{
The rows of a input numeric matrix is ordered based row sum, row maximum, or hierarchical clustering of the rows with euclidean
distannce and centroid linkage.
}
\author{
Shuye Pu
}
