% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDataMatrix.R
\name{ratio_over_input}
\alias{ratio_over_input}
\title{compute ratio over input}
\usage{
ratio_over_input(IP, Input, verbose = FALSE)
}
\arguments{
\item{IP}{a numerical matrix}

\item{Input}{another numerical matrix with same dimensions as the IP matrix}

\item{verbose}{logical, whether to output additional information}
}
\value{
a numerical matrix with same dimensions as the IP matrix
}
\description{
compute enrichment of IP samples over Input samples
}
\author{
Shuye Pu
}
\keyword{internal}
