% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{rmOutlier}
\alias{rmOutlier}
\title{rmOutlier}
\usage{
rmOutlier(fullmatrix)
}
\arguments{
\item{fullmatrix, }{a numeric matrix, with bins in columns and genomic windows in rows}
}
\value{
a numeric matrix
}
\description{
This is a helper function for removing rows with excessively high values, using Hampel filter with 10mad instead of 3mad.
If outliers are detected, remove the top outlier only, to minimize data loss.
}
