% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleFeatures.R
\name{set_seqinfo}
\alias{set_seqinfo}
\title{Set standard chromosome size of model organisms}
\usage{
set_seqinfo(genome = "hg19")
}
\arguments{
\item{genome}{a string denoting the genome name and version}
}
\value{
a Seqinfo object defined in the Seqinfo package.
}
\description{
This is a helper function for making Seqinfo objects, which is
a components of GRanges and TxDb objects. It also serves to unify
seqlevels between GRanges and TxDb objects. Mitochondrial chromosome is not
included.
}
\examples{

out <- set_seqinfo(genome = "hg19")

}
\author{
Shuye Pu
}
