% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{start_parallel}
\alias{start_parallel}
\title{start_parallel}
\usage{
start_parallel(nc)
}
\arguments{
\item{nc, }{a positive integer greater than 1}
}
\value{
an object of class c("SOCKcluster", "cluster"), depending on platform
}
\description{
Method for starting a virtual cluster needed for parallel processing
}
\examples{
cl <- start_parallel(5L)

}
