% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{stop_parallel}
\alias{stop_parallel}
\title{stop_parallel
Method for stopping a virtual cluster needed for parallel processing}
\usage{
stop_parallel(cl)
}
\arguments{
\item{cl, }{a cluster or SOCKcluster object depending on platform}
}
\description{
stop_parallel
Method for stopping a virtual cluster needed for parallel processing
}
\examples{
cl <- start_parallel(5L)
stop_parallel(cl)

}
