% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb.sql}
\alias{txdb.sql}
\title{Toy data for examples and testing of the `GenomicPlot` package}
\source{
The data is produced by running the following code: \cr
gtffile <- system.file("extdata", "gencode.v19.annotation_chr19.gtf", 
package = "GenomicPlot") \cr
gff <- RCAS::importGtf(saveObjectAsRds = TRUE, readFromRds = FALSE, 
    filePath = gtffile) \cr
md <- data.frame(name = "Genome", value = "hg19") \cr
txdb <- GenomicFeatures::makeTxDbFromGRanges(gff, metadata = md) \cr
AnnotationDbi::saveDb(txdb, "./inst/extdata/txdb.sql")
}
\value{
A SQLlite database
}
\description{
A tiny TxDb object holding genomic feature coordinates of 72 
transcripts in hg19.
}
\author{
Shuye Pu
}
\keyword{datasets}
