% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleFeatures.R
\name{check_constraints}
\alias{check_constraints}
\title{Check constraints of genomic ranges}
\usage{
check_constraints(gr, genome, queryRle = NULL)
}
\arguments{
\item{gr}{a GenomicRanges object}

\item{genome}{genomic version name such as "hg18"}

\item{queryRle}{a RleList object used as a query against gr}
}
\value{
a GRanges object
}
\description{
Make sure the coordinates of GRanges are within the boundaries of chromosomes, and trim anything that goes beyound. Also, remove entries whose seqname is not the seqname of a query GRanges.
}
\author{
Shuye Pu
}
