% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawingFunctions.R
\name{draw_region_profile}
\alias{draw_region_profile}
\title{Plot signal profile in genomic regions}
\usage{
draw_region_profile(
  plot_df,
  xc = "Position",
  yc = "Intensity",
  cn = "Query",
  sn = "Reference",
  Ylab = "Signal Intensity",
  vx
)
}
\arguments{
\item{plot_df}{a dataframe with column names c(xc, yc, cn, "lower", "upper")}

\item{xc}{a string denoting column name for values on x-axis}

\item{yc}{a string denoting column name for numeric data to be plotted}

\item{cn}{column name in plot_df for query samples grouping}

\item{sn}{column name in plot_df for subject name to be shown in the plot title}

\item{Ylab}{a string for Y-axis label}

\item{vx}{a vector on integers denoting the x coordinates of start of each sub-region}
}
\value{
a ggplot object
}
\description{
Plot lines with standard error as the error band
}
\note{
used by \code{plot_3parts_metagene}, \code{plot_5parts_metagene}, \code{plot_region}
}
\author{
Shuye Pu
}
