% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{effective_size}
\alias{effective_size}
\title{Normalize sample library size to effective size}
\usage{
effective_size(outlist, outRle, genome = "hg19", nc = 2)
}
\arguments{
\item{outlist}{a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type,
'weight' is the name of the metadata column}

\item{outRle}{logical, indicating whether the output is a list of RleList objects or GRanges objects}

\item{genome}{a string denoting the genome name and version}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a list object with four elements ('query', 'size', 'type', 'weight'), with the 'size' element modified.
}
\description{
This is a helper function for handle_input. edgeR::calcNormFactors function is used to estimate normalizing factors,
which is used to multiply library sizes. The function only works for human genome only at present.
}
\author{
Shuye Pu
}
