% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{filter_by_nonoverlaps_stranded}
\alias{filter_by_nonoverlaps_stranded}
\title{Filter GRanges by nonoverlaps in stranded way}
\usage{
filter_by_nonoverlaps_stranded(query, subject)
}
\arguments{
\item{query}{a GRanges object}

\item{subject}{a GRanges object}
}
\value{
a GRanges object
}
\description{
This function reports all query GRanges that do not overlaps GRanges in subject. Strand information is used to define overlap.
}
\author{
Shuye Pu
}
