% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleFeatures.R
\name{gene2tx}
\alias{gene2tx}
\title{Translate gene names to transcript ids using a GTF file for a subset of genes}
\usage{
gene2tx(gtfFile, geneList, geneCol = 1)
}
\arguments{
\item{geneList}{path to a line-delimited text file with one gene name on each line, or a character vector of gene names}

\item{gtfFfile}{path to a GTF file}
}
\value{
a vector
}
\description{
Given a list of gene names in a file or in a character vector, turn them into a vector of transcript ids.
}
\author{
Shuye Pu
}
