% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{handle_bed}
\alias{handle_bed}
\title{Handle files in bed format}
\usage{
handle_bed(inputFile, handleInputParams = NULL)
}
\arguments{
\item{inputFile}{a string denoting path to the input file}

\item{handleInputParams}{a list of parameters for \code{handle_input}}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type, 'weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
This is a function for read peaks data in bed format, store the input data in a list of GRanges objects or RleList objects.
}
\author{
Shuye Pu
}
