% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{handle_bed}
\alias{handle_bed}
\title{Handle files in bed|narrowPeak|broadPeak format}
\usage{
handle_bed(inputFile, importParams = NULL, verbose = FALSE)
}
\arguments{
\item{inputFile}{a string denoting path to the input file}

\item{importParams}{a list of parameters, refer to \code{handle_input} for details}

\item{verbose}{logical, whether to output additional information}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type, 'weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
This is a function for read peaks data in bed format, store the input data in a list of GRanges objects or RleList objects.
}
\author{
Shuye Pu
}
