% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{handle_bw}
\alias{handle_bw}
\title{Handle files in bw|bigwig|bigWig|BigWig|BW|BIGWIG format}
\usage{
handle_bw(inputFile, handleInputParams, verbose = FALSE)
}
\arguments{
\item{inputFile}{a string denoting path to the input file}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, whether to output additional information}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the estimated library size, 'type' is the input file type, weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
This is a function for read NGS coverage data in bigwig format, store the input data in a list of GRanges objects or RleList objects. The input bw file can be stranded or non-stranded. Library size is calculate as the sum of all coverage.
}
\details{
For stranded files, forward and reverse strands are stored in separate files, with '+' or 'p' in the forward strand file name and '-' or 'm' in the reverse strand  file name.
}
\author{
Shuye Pu
}
