% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{handle_input}
\alias{handle_input}
\title{Hand input of NGS data with various formats}
\usage{
handle_input(inputFiles, handleInputParams = NULL, verbose = FALSE, nc = 2)
}
\arguments{
\item{inputFiles}{a vector of strings denoting file names}

\item{handleInputParams}{a list with the following elements:
'offset' integer, -1 indicating the bam reads should be shrunk to the -1 position at the 5'end of the reads, which corresponds to the cross link site in iCLIP
'fix_width' an integer defines how long should the reads should be extended to.
'fix_point' a string in c("start", "end", "center") denoting the anchor point for extension.
'useScore' logical, indicating whether the 'score' column of the bed file should be used in calculation of coverage.
'outRle' logical, indicating whether the output should be RleList objects or GRanges objects.
'norm' logical, indicating whether the output RleList should be normalized to RPM using library sizes.
'genome' a string denoting the genome name and version.
'useSizeFactor' logical, indicating whether the library size should be adjusted with a size factor, using the 'calcNormFactors' function in the edgeR package}

\item{verbose}{logical, whether to output additional information}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a list object with four elements, 'query' is a list GRanges objects or RleList objects, 'size' is the library size, 'type' is the input file type,
'weight' is the name of the metadata column to be used as weight for coverage calculation
}
\description{
This is a wrapper function for read NGS data in different file formats, store the input data in a list of GRanges objects or RleList objects. File names end in bed|bam|bw|bigwig|bigWig|BigWig|BW|BIGWIG are recognized, and a list of files with mixed formats are allowed.
}
\details{
when 'useScore' is TRUE, the score column of the bed file will be used in the metadata column 'score' of the GRanges object, or the 'Values' field of the RleList object. Otherwise the value 1 will be used instead. When the intended use of the input bed is a reference feature, both 'useScore' and 'outRle' should be set to FALSE.
}
\examples{
queryFiles <- system.file("extdata", "treat_chr19.bam", package="GenomicPlot")
names(queryFiles) <- "query"

inputFiles <- system.file("extdata", "input_chr19.bam", package="GenomicPlot")
names(inputFiles) <- "input"

handleInputParams <- list(offset=-1, fix_width=0, fix_point="start", norm=TRUE, 
useScore=FALSE, outRle=TRUE, useSizeFactor=TRUE, genome="hg19")

out_list <- handle_input(inputFiles=c(queryFiles, inputFiles), 
handleInputParams=handleInputParams, verbose=TRUE, nc=2)

}
\author{
Shuye Pu
}
