% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDataMatrix.R
\name{impute_hm}
\alias{impute_hm}
\title{Impute missing values}
\usage{
impute_hm(fullmatrix, verbose = FALSE)
}
\arguments{
\item{fullmatrix}{a numeric matrix}

\item{verbose}{logical, whether to output additional information}
}
\value{
a numeric matrix
}
\description{
Replace 0 and missing values in a matrix with half of minimum, to avoid use of arbitrary pseudo numbers
and to allow log transformation
}
\keyword{internal}
