% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDataMatrix.R
\name{inspect_matrix}
\alias{inspect_matrix}
\title{Inspect a numeric matrix}
\usage{
inspect_matrix(fullmatrix, verbose = FALSE)
}
\arguments{
\item{fullmatrix}{a numeric matrix}

\item{verbose}{logical, indicating whether to print out the stats in the console}
}
\description{
Check the matrix for NA, NaN, INF, -INF and 0 values
}
\keyword{internal}
