% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_3parts_metagene}
\alias{plot_3parts_metagene}
\title{Plot promoter, gene body and TTS}
\usage{
plot_3parts_metagene(
  queryFiles,
  gFeatures,
  inputFiles = NULL,
  scale = FALSE,
  verbose = FALSE,
  Ylab = "Coverage/base/gene",
  importParams = NULL,
  smooth = FALSE,
  stranded = TRUE,
  outPrefix = NULL,
  heatmap = FALSE,
  rmOutlier = 0,
  heatRange = NULL,
  transform = NA,
  hw = c(8, 8),
  nc = 2
)
}
\arguments{
\item{queryFiles}{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{gFeatures}{genomic features as output of the function 'prepare_3parts_genomic_features'}

\item{inputFiles}{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{verbose}{logical, whether to output additional information (data used for plotting or statistical test results)}

\item{Ylab}{a string for y-axis label}

\item{importParams}{a list of parameters for \code{handle_input}}

\item{smooth}{logical, indicating whether the line should smoothed with a spline smoothing algorithm}

\item{stranded}{logical, indicating whether the strand of the feature should be considered}

\item{outPrefix}{a string specifying output file prefix for plots (outPrefix.pdf)}

\item{heatmap}{logical, indicating whether a heatmap of the score matrix should be generated}

\item{rmOutlier}{a numeric value serving as a multiplier of the MAD in Hampel filter for outliers identification, 0 indicating not removing outliers. For Gaussian distribution, use 3, adjust based on data distribution}

\item{heatRange}{a numerical vector of two elements, defining range for heatmap color ramp generation}

\item{transform}{a string in c("log", "log2", "log10"), default = NA indicating no transformation of data matrix}

\item{hw}{a vector of two elements specifying the height and width of the output figures}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a dataframe containing the data used for plotting
}
\description{
Plot reads or peak Coverage/base/gene of samples given in the query files around genes. The upstream and downstream windows flanking genes
can be given separately, the parameter 'meta' controls if gene or metagene plots are generated. If Input files are provided, ratio over Input
is computed and displayed as well.
}
\author{
Shuye Pu
}
