% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_3parts_metagene}
\alias{plot_3parts_metagene}
\title{Plot promoter, gene body and TTS}
\usage{
plot_3parts_metagene(
  queryFiles,
  gFeatures,
  inputFiles = NULL,
  scale = FALSE,
  verbose = FALSE,
  Ylab = "Signal intensity",
  handleInputParams = NULL,
  smooth = FALSE,
  stranded = TRUE,
  outPrefix = "plots",
  heatmap = FALSE,
  rmOutlier = FALSE,
  heatRange = NULL,
  transform = FALSE,
  nc = 2
)
}
\arguments{
\item{queryFiles}{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{gFeatures}{genomic features as output of the function 'prepare_3parts_genomic_features'}

\item{inputFiles}{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{verbose}{logical, whether to output additional information (data used for plotting or statistical test results)}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{smooth}{logical, indicating whether the line should smoothed with a spline smoothing algorithm}

\item{stranded}{logical, indicating whether the strand of the feature should be considered}

\item{outPrefix}{a string specifying output file prefix for plots (outPrefix.pdf)}

\item{heatmap}{logical, indicating whether a heatmap of the score matrix should be generated}

\item{rmOutlier}{logical, indicating whether a row with abnormally high values in the score matrix should be removed}

\item{heatRange}{a numerical vector of two elements, defining range for heatmap color ramp generation}

\item{transform}{logical, whether to log2 transform the data matrix}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a dataframe containing the data used for plotting
}
\description{
Plot reads or peak signal intensity of samples in the query files around genes. The upstream and downstream windows flanking genes
can be given separately, the parameter 'meta' controls if gene or metagene plots are generated. If Input files are provided, ratio over Input
is computed and displaued as well.
}
\examples{
centerFiles <- c(system.file("data", "test_B.bed", package="GenomicPlotData"),
system.file("data", "test_C.bed", package="GenomicPlotData"))
names(centerFiles) <- c("TestB", "TestC")

}
\author{
Shuye Pu
}
