% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_5parts_metagene}
\alias{plot_5parts_metagene}
\title{Plot promoter, 5'UTR, CDS, 3'UTR and TTS}
\usage{
plot_5parts_metagene(
  queryfiles,
  gFeatures_list,
  inputfiles = NULL,
  handleInputParams = NULL,
  verbose = FALSE,
  transform = FALSE,
  smooth = FALSE,
  scale = FALSE,
  stranded = TRUE,
  outPrefix = NULL,
  heatmap = FALSE,
  rmOutlier = FALSE,
  nc = 2
)
}
\arguments{
\item{queryfiles}{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{gFeatures_list}{a list of genomic features as output of the function 'prepare_5parts_genomic_features'}

\item{inputfiles}{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, indicating whether to output additional information (data used for plotting or statistical test results)}

\item{transform}{logical, whether to log2 transform the matrix}

\item{smooth}{logical, indicating whether the line should smoothed with a spline smoothing algorithm}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{stranded}{logical, indicating whether the strand of the feature should be considered}

\item{outPrefix}{a string specifying output file prefix for plots (outPrefix.pdf)}

\item{heatmap}{logical, indicating whether a heatmap of the score matrix should be generated}

\item{rmOutlier}{logical, indicating whether a row with abnormally high values in the score matrix should be removed}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a dataframe containing the data used for plotting
}
\description{
Plot reads or peak signal intensity of samples in the query files around genes. The upstream and downstream windows flanking genes can be given separately, metagene plots are generated with 5'UTR, CDS and 3'UTR segments. The length of each segments are prorated according to the median length of each segments. If Input files are provided, ratio over Input is computed and displayed as well.
}
\examples{
txdb <- AnnotationDbi::loadDb(system.file("data", "txdb_chr19.sql", package="GenomicPlotData"))
gf <- prepare_5parts_genomic_features(txdb, meta=FALSE, nbins=100, fiveP=500, threeP=500, longest=TRUE)
queryfiles <- c(system.file("data", "test_chip_peak_chr19.narrowPeak", package="GenomicPlotData"),
system.file("data", "test_chip_peak_chr19.bed", package="GenomicPlotData"),
system.file("data", "test_clip_peak_chr19.bed", package="GenomicPlotData"))
names(queryfiles) <- c("narrowPeak", "summitPeak", "iCLIPPeak")
op <- NULL

handleInputParams <- list(CLIP_reads=FALSE, fix_width=150, fix_point="center", norm=FALSE, useScore=FALSE,
                          outRle=TRUE, useSizeFactor=FALSE, genome="hg19")

plot_5parts_metagene(queryfiles, gFeatures=gf, inputfiles=NULL, handleInputParams=handleInputParams,
                     verbose=FALSE, smooth=TRUE, scale=FALSE, stranded=TRUE, outPrefix=op, transform=FALSE,
                     heatmap =TRUE,  rmOutlier=TRUE, nc=2)

}
\author{
Shuye Pu
}
