% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_bam_correlation}
\alias{plot_bam_correlation}
\title{Plot correlation of bam files}
\usage{
plot_bam_correlation(
  bamfiles,
  binSize = 1e+06,
  outPrefix = NULL,
  importParams = NULL,
  verbose = FALSE,
  hw = c(8, 8),
  nc = 2
)
}
\arguments{
\item{bamfiles}{a named vector of strings denoting file names}

\item{binSize}{an integer denoting the tile width for tiling the genome, default 1000000}

\item{outPrefix}{a string denoting output file name in pdf format}

\item{importParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, indicating whether to output additional information}

\item{hw}{a vector of two elements specifying the height and width of the output figures}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a dataframe of read counts per bin perl sample
}
\description{
plot correlation in reads coverage distributions along the genome for bam files
}
\examples{
# Due to file size limitation, peak files are used instead.

bedQueryFiles <- c(
  system.file("extdata", "test_chip_peak_chr19.narrowPeak", package = "GenomicPlot"),
  system.file("extdata", "test_chip_peak_chr19.bed", package = "GenomicPlot"),
  system.file("extdata", "test_clip_peak_chr19.bed", package = "GenomicPlot")
)
names(bedQueryFiles) <- c("NarrowPeak", "SummitPeak", "iCLIPPeak")

bedimportParams <- list(
   offset = 0, fix_width = 100, fix_point = "center", norm = FALSE,
   useScore = FALSE, outRle = FALSE, useSizeFactor = FALSE, genome = "hg19"
)

plot_bam_correlation(
  bamfiles = bedQueryFiles, binSize = 10000, outPrefix = NULL,
  importParams = bedimportParams, nc = 2
)

}
