% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_bam_correlation}
\alias{plot_bam_correlation}
\title{Plot correlation of bam files}
\usage{
plot_bam_correlation(
  bamfiles,
  binSize = 1e+06,
  outPrefix = NULL,
  handleInputParams = NULL,
  verbose = FALSE,
  nc = 2
)
}
\arguments{
\item{bamfiles}{a named vector of strings denoting file names}

\item{binSize}{an integer denoting the tile width for tiling the genome, default 1000000}

\item{outPrefix}{a string denoting output file name in pdf format}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, indicating whether to output additional information}

\item{nc}{integer, number of cores for parallel processing}
}
\description{
plot correlation in reads coverage distributions along the genome for bam files
}
\examples{
queryFiles <- c(system.file("data", "test_chip_chr19.bam", package="GenomicPlotData"),
system.file("data", "test_clip_chr19.bam", package="GenomicPlotData"))
names(queryFiles) <- c("chip_query", "clip_query")
inputFiles <- c(system.file("data", "test_chip_input_chr19.bam", package="GenomicPlotData"),
                system.file("data", "test_clip_input_chr19.bam", package="GenomicPlotData"))
names(inputFiles) <- c("chip_input", "clip_input")

op=NULL
handleInputParams <- list(CLIP_reads=FALSE, fix_width=0, fix_point="start", norm=FALSE, useScore=FALSE,
                          outRle=FALSE, useSizeFactor=FALSE, genome="hg19")
plot_bam_correlation(bamfiles=c(queryFiles[1], inputFiles[1]), binSize=10000, outPrefix=op,
                     handleInputParams=handleInputParams, nc=2)


}
