% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_bam_correlation}
\alias{plot_bam_correlation}
\title{Plot correlation of bam files}
\usage{
plot_bam_correlation(
  bamfiles,
  binSize = 1e+06,
  outPrefix = NULL,
  handleInputParams = NULL,
  verbose = FALSE,
  wh = c(8, 8),
  nc = 2
)
}
\arguments{
\item{bamfiles}{a named vector of strings denoting file names}

\item{binSize}{an integer denoting the tile width for tiling the genome, default 1000000}

\item{outPrefix}{a string denoting output file name in pdf format}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, indicating whether to output additional information}

\item{nc}{integer, number of cores for parallel processing}

\item{hw}{a vector of two elements specifying the height and width of the output figures}
}
\description{
plot correlation in reads coverage distributions along the genome for bam files
}
\examples{
queryFiles <- c(system.file("extdata", "chip_treat_chr19.bam", package="GenomicPlot"),
               system.file("extdata", "treat_chr19.bam", package="GenomicPlot"))
names(queryFiles) <- c("chip_query", "clip_query")

inputFiles <- c(system.file("extdata", "chip_input_chr19.bam", package="GenomicPlot"),
                system.file("extdata", "input_chr19.bam", package="GenomicPlot"))
names(inputFiles) <- c("chip_input", "clip_input")

handleInputParams <- list(CLIP_reads=FALSE, fix_width=0, fix_point="start", norm=FALSE, 
useScore=FALSE, outRle=FALSE, useSizeFactor=FALSE, genome="hg19")
                          
plot_bam_correlation(bamfiles=c(queryFiles, inputFiles), binSize=10000, outPrefix=NULL,
   handleInputParams=handleInputParams, nc=2)


}
