% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_overlap_genes}
\alias{plot_overlap_genes}
\title{Plot Venn diagrams depicting overlap of gene lists}
\usage{
plot_overlap_genes(
  fileList,
  columnList,
  pairOnly = TRUE,
  hw = c(8, 8),
  outPrefix = NULL
)
}
\arguments{
\item{fileList, }{a named list of tab-dlimited files}

\item{columnList}{a vector of integers denoting the columns that have gene names in the list of files}

\item{pairOnly, }{logical, indicating whether only pair-wise overlap is desired}

\item{hw}{a vector of two elements specifying the height and width of the output figures}

\item{outPrefix, }{a string for plot file name}
}
\value{
a list of vectors of gene names
}
\description{
This function takes a list of (at most 3) tab-delimited file names, and produce a Venn diagram
}
\author{
Shuye Pu
}
