% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_reference_locus_with_random}
\alias{plot_reference_locus_with_random}
\title{Plot signal around custom genomic loci and random loci for comparison}
\usage{
plot_reference_locus_with_random(
  queryFiles,
  centerFiles,
  txdb,
  ext = c(0, 0),
  hl = c(0, 0),
  shade = FALSE,
  handleInputParams = NULL,
  verbose = FALSE,
  smooth = FALSE,
  transform = FALSE,
  binSize = 10,
  refPoint = "center",
  Xlab = "Center",
  inputFiles = NULL,
  stranded = TRUE,
  scale = FALSE,
  outPrefix = NULL,
  rmOutlier = FALSE,
  n_random = 1,
  statsMethod = "wilcox.test",
  nc = 2
)
}
\arguments{
\item{queryFiles}{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{centerFiles}{a vector of reference file names. The file should be .bed format only}

\item{txdb}{a TxDb object defined in the GenomicFeatures package}

\item{ext}{a vector of two integers defining upstream and downstream boundaries of the plot window, flanking the reference locus}

\item{hl}{a vector of two integers defining upstream and downstream boundaries of the highlight window, flanking the reference locus}

\item{shade}{logical indicating whether to place a shaded rectangle around the point of interest}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, indicating whether to output additional information (data used for plotting or statistical test results)}

\item{smooth}{logical, indicating whether the line should smoothed with a spline smoothing algorithm}

\item{transform}{logical, whether to log2 transform the data matrix}

\item{binSize}{an integer defines bin size for intensity calculation}

\item{refPoint}{a string in c("start", "center", "end")}

\item{Xlab}{a string denotes the label on x-axis}

\item{inputFiles}{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{stranded}{logical, indicating whether the strand of the feature should be considered}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{outPrefix}{a string specifying output file prefix for plots (outPrefix.pdf)}

\item{rmOutlier}{logical, indicating whether a row with abnormally high values in the score matrix should be removed}

\item{n_random}{an integer denotes the number of randomization should be formed}

\item{statsMethod}{a string in c("wilcox.test", "t.test"), for pair-wise groups comparisons}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a dataframe containing the data used for plotting
}
\description{
Plot reads or peak signal intensity of samples in the query files around reference locus defined in the centerFiles. The upstream and downstream windows flanking genes
can be given separately, a smaller window can be defined to allow statistical comparisons between reference and random loci.
}
\examples{
queryFiles <- system.file("data", "test_clip.bam", package="GenomicPlotData")
names(queryFiles) <- "query"
inputFiles <- system.file("data", "test_clip_input.bam", package="GenomicPlotData")
names(inputFiles) <- "input"

}
\author{
Shuye Pu
}
