% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_reference_region}
\alias{plot_reference_region}
\title{Plot signal inside as well as around custom genomic regions}
\usage{
plot_reference_region(
  queryfiles,
  centerfiles,
  inputfiles = NULL,
  nbins = 100,
  handleInputParams = NULL,
  verbose = FALSE,
  scale = FALSE,
  heatmap = FALSE,
  regionName = "region",
  fiveP = 1000,
  threeP = 1000,
  smooth = FALSE,
  stranded = TRUE,
  transform = FALSE,
  outPrefix = "plots",
  rmOutlier = FALSE,
  nc = 2
)
}
\arguments{
\item{queryfiles}{a named vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{centerfiles}{a named vector of reference file names. The file should be .bed format only}

\item{inputfiles}{a named vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{nbins}{an integer defines the total number of bins}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{verbose}{logical, indicating whether to output additional information (data used for plotting or statistical test results)}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{heatmap}{logical, indicating whether a heatmap of the score matrix should be generated}

\item{regionName}{a string specifying the name of the center region in the plots}

\item{fiveP}{extension out of the 5' boundary of gene}

\item{threeP}{extension out of the 3' boundary of gene}

\item{smooth}{logical, indicating whether the line should smoothed with a spline smoothing algorithm}

\item{stranded}{logical, indicating whether the strand of the feature should be considered}

\item{transform}{logical, whether to log2 transform the data matrix}

\item{outPrefix}{a string specifying output file prefix for plots (outPrefix.pdf)}

\item{rmOutlier}{logical, indicating whether a row with abnormally high values in the score matrix should be removed}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a dataframe containing the data used for plotting
}
\description{
Plot reads or peak signal intensity of samples in the query files inside regions defined in the centerfiles. The upstream and downstream flanking windows
can be given separately. If Input files are provided, ratio over Input is computed and displaued as well.
}
\examples{
centerfiles <- c(system.file("data", "test_B.bed", package="GenomicPlotData"),
system.file("data", "test_C.bed", package="GenomicPlotData"))
names(centerfiles) <- c("TestB", "TestC")

}
\author{
Shuye Pu
}
