% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicPlot.R
\name{plot_start_end}
\alias{plot_start_end}
\title{Plot signals around the start and the end of genomic features}
\usage{
plot_start_end(
  queryFiles,
  inputFiles = NULL,
  centerFiles,
  txdb = NULL,
  handleInputParams = NULL,
  binSize = 10,
  insert = 0,
  verbose = FALSE,
  ext = c(-500, 100, -100, 500),
  hl = c(-50, 50, -50, 50),
  stranded = TRUE,
  scale = FALSE,
  smooth = FALSE,
  rmOutlier = FALSE,
  outPrefix = "plots",
  transform = NA,
  shade = TRUE,
  Ylab = "Coverage/base/gene",
  nc = 2
)
}
\arguments{
\item{queryFiles}{a vector of sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{inputFiles}{a vector of input sample file names. The file should be in .bam, .bed, .wig or .bw format, mixture of formats is allowed}

\item{centerFiles}{bed files that define the custom features, or features in c("utr3", "utr5", "cds", "intron", "exon", "transcript", "gene"), multiple features are allowed.}

\item{txdb}{a TxDb object defined in the GenomicFeatures package. Default NULL, needed only when genomic features are used in the place of centerFiles.}

\item{handleInputParams}{a list of parameters for \code{handle_input}}

\item{binSize}{an integer defines bin size for intensity calculation}

\item{insert}{an integer specifies the length of the center regions to be included, in addition to the start and end of the feature}

\item{verbose}{logical, whether to output additional information (including data used for plotting or statistical test results)}

\item{ext}{a vector of four integers defining upstream and downstream boundaries of the plot window, flanking the start and end of features}

\item{hl}{a vector of four integers defining upstream and downstream boundaries of the highlight window, flanking the start and end of features}

\item{stranded}{logical, indicating whether the strand of the feature should be considered}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{smooth}{logical, indicating whether the line should smoothed with a spline smoothing algorithm}

\item{rmOutlier}{logical, indicating whether a row with abnormally high values in the score matrix should be removed}

\item{outPrefix}{a string specifying output file prefix for plots (outPrefix.pdf)}

\item{transform}{a string in c("log", "log2", "log10"), default = NA, indicating no transformation of data matrix}

\item{shade}{logical indicating whether to place a shaded rectangle around the point of interest}

\item{Ylab}{a string for y-axis label}

\item{nc}{integer, number of cores for parallel processing}
}
\value{
a list of two objects, the first is a GRanges object, the second is a GRangesList object
}
\description{
Plot reads or peak Coverage/base/gene of samples in the query files around start and end of custom features. The upstream and downstream windows
can be given separately, within the window, a smaller window can be defined to highlight region of interest.
A line plot will be displayed for both start and end of feature. If Input files are provided, ratio over Input is
computed and displayed as well.
}
\examples{
txdb <- AnnotationDbi::loadDb(system.file("extdata", "txdb_chr19.sql", package="GenomicPlot"))
queryfiles <- system.file("extdata", "treat_chr19.bam", package="GenomicPlot")
names(queryfiles) <- "query"
inputfiles <- system.file("extdata", "input_chr19.bam", package="GenomicPlot")
names(inputfiles) <- "input"
centerfiles <- system.file("extdata", "test_chip_peak_chr19.narrowPeak", package="GenomicPlot")
names(centerfiles) <- "narrowPeak"
op <- NULL
handleInputParams <- list(CLIP_reads=FALSE, fix_width=150, fix_point="start", norm=FALSE, useScore=FALSE,
 outRle=TRUE, useSizeFactor=TRUE, genome="hg19")
 
df <- plot_start_end(queryFiles=queryfiles, inputFiles=inputfiles, centerFiles=c("gene", centerfiles), txdb=txdb, 
handleInputParams=handleInputParams, binSize=10, insert=100, verbose=TRUE, ext=c(-500, 100, -100, 500), 
hl=c(-50, 50, -50, 50), stranded=TRUE, scale=FALSE, smooth=TRUE, rmOutlier=FALSE, outPrefix=op, 
transform=NA, shade=TRUE, Ylab="Coverage/base/gene", nc=2)
 
}
\author{
Shuye Pu
}
