% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleDataMatrix.R
\name{process_scoreMatrix}
\alias{process_scoreMatrix}
\title{Preprocess scoreMatrix before plotting}
\usage{
process_scoreMatrix(
  fullmatrix,
  scale = FALSE,
  rmOutlier = FALSE,
  transform = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{fullmatrix}{a numeric matrix, with bins in columns and genomic windows in rows}

\item{scale}{logical, indicating whether the score matrix should be scaled to the range 0:1, so that samples with different baseline can be compared}

\item{rmOutlier}{logical, indicating whether a row with abnormally high values in the score matrix should be removed}

\item{transform}{logical, whether to log2 transform the data matrix}

\item{verbose}{logical, indicating whether to output additional information (data used for plotting or statistical test results)}
}
\value{
a numeric matrix with the same dimension as the fullmatrix
}
\description{
This is a helper function for manipulate the score matrix produced by ScoreMatrix or ScoreMatrinBin functions defined in the 'genomation' package.
}
\author{
Shuye Pu
}
