% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandleFeatures.R
\name{set_seqinfo}
\alias{set_seqinfo}
\title{Set standard chromosome size of model organisms}
\usage{
set_seqinfo(genome = "hg19", chromInfo = NULL)
}
\arguments{
\item{genome}{a string denoting the genome name and version}

\item{chromInfo}{a data frame with three columns: chr, start and end}
}
\value{
a Seqinfo object defined in the Seqinfo package.
}
\description{
This is a helper function for making Seqinfo objects, which is
a components of GRanges and TxDb objects. It also serves to unify
seqlevels between GRanges and TxDb objects. Mitochondrial chromosome is not
included.
}
\examples{

# Default usage with standard chromosome information
out <- set_seqinfo(genome = "hg19")

# Custom chromosome information
custom_chrom <- data.frame(chr=c("chr1","chr2"), start=c(0,0), end=c(1000000,2000000))
out_custom <- set_seqinfo(genome = "hg199", chromInfo = custom_chrom)

}
\author{
Shuye Pu
}
