% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{start_parallel}
\alias{start_parallel}
\title{Prepare for parallel processing}
\usage{
start_parallel(nc = 2, verbose = FALSE)
}
\arguments{
\item{nc}{a positive integer greater than 1, denoting number of cores requested}

\item{verbose}{logical, whether to output additional information}
}
\value{
an object of class c("SOCKcluster", "cluster"), depending on platform
}
\description{
Creating a virtual cluster for parallel processing
}
\examples{

cl <- start_parallel(2L)
stop_parallel(cl)

}
\author{
Shuye Pu
}
