% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{stop_parallel}
\alias{stop_parallel}
\title{Stop parallel processing}
\usage{
stop_parallel(cl)
}
\arguments{
\item{cl}{a cluster or SOCKcluster object depending on platform}
}
\value{
0 if the cluster is stopped successfully, 1 otherwise.
}
\description{
Stopping a virtual cluster after parallel processing is finished
}
\examples{
cl <- start_parallel(2L)
stop_parallel(cl)

}
\author{
Shuye Pu
}
