% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb.sql}
\alias{txdb.sql}
\title{Toy data for examples and testing of the `GenomicPlot` package}
\format{
SQLlite database
}
\source{
The data is produced by running the following code:
\describe{
gtffile <- system.file("extdata", "gencode.v19.annotation_chr19.gtf", 
package = "GenomicPlot")
}
\describe{
gff <- RCAS::importGtf(saveObjectAsRds = TRUE, readFromRds = FALSE, 
    filePath = gtffile)
}
\describe{
md <- data.frame(name = "Genome", value = "hg19")
}
\describe{
txdb <- GenomicFeatures::makeTxDbFromGRanges(gff, metadata = md)
}
\describe{
AnnotationDbi::saveDb(txdb, "./inst/extdata/txdb.sql")
}
}
\description{
A tiny TxDb object holding genomic feature coordinates of 72 
transcripts in hg19.
}
\author{
Shuye Pu
}
\keyword{datasets}
